Refactor test structure depending on modules
